Move TimePicker to date picker folder
